requirements.txt
fastapi
uvicorn
pydantic
starlette
click
anyio
sniffio
websockets
python-multipart
twilio
python-dotenv
pydub
python-dateutil
six
httpx
pygments
rich
attrs
openai
PyJWT
tzdata
chromadb
langchain-openai
langchain-text-splitters
google-genai>=1.73.0
resemblyzer
numpy
